+++ src/Vector3.cpp
@@ -8,7 +8,7 @@
 #include <cmath>
 #include <sstream>
 
-Vector3::Vector3(float x, float y, float z):
+Vector3::Vector3(double x, double y, double z):
     x(x), y(y), z(z){
 }
 
@@ -26,15 +26,15 @@ Vector3 Vector3::sub(Vector3 other){
 	return Vector3(this->x - other.x, this->y - other.y, this->z - other.z);
 }
 
-Vector3 Vector3::mul(float k){
+Vector3 Vector3::mul(double k){
 	return Vector3(this->x * k, this->y * k, this->z * k);
 }
 
-float Vector3::dot(Vector3 other){
+double Vector3::dot(Vector3 other){
 	return this->x * other.x + this->y*other.y + this->z * other.z;
 }
 
-float Vector3::length(){
+double Vector3::length(){
 	return std::sqrt(this->dot(*this));
 }
